requirements.txt
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
onnx==1.19.1
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
triton==3.3.1
typing_extensions==4.15.0
tzdata==2025.2
